requirements.txt
streamlit
pandas
huggingface_hub
python-dotenv
supabase
plotly
openpyxl